Refactor get code into the same service
