+++ NCS.DSS.DigitalIdentity/GetDigitalIdentityHttpTrigger/Service/GetDigitalIdentityHttpTriggerService.cs
@@ -14,6 +14,24 @@ namespace NCS.DSS.DigitalIdentity.GetDigitalIdentityHttpTrigger.Service
         {
             _documentDbProvider = documentDbProvider;
         }
+        public async Task<Models.DigitalIdentity> GetIdentityForCustomerAsync(Guid customerId)
+        {
+            var identity = await _documentDbProvider.GetIdentityForCustomerAsync(customerId);
+
+            return identity;
+        }
+
+        public async Task<Models.DigitalIdentity> GetIdentityByIdentityIdAsync(Guid identityGuid)
+        {
+            var identity = await _documentDbProvider.GetIdentityByIdentityIdAsync(identityGuid);
+
+            return identity;
+        }
+
+        public async Task<bool> DoesCustomerExists(Guid? identityRequestCustomerId)
+        {
+            return identityRequestCustomerId != null && await _documentDbProvider.DoesCustomerResourceExist(identityRequestCustomerId.Value);
+        }
 
         public async Task<Models.DigitalIdentity> GetIdentityAsync(Guid identityId)
         {
