--- NCS.DSS.DigitalIdentity/GetDigitalIdentityHttpTrigger/function.json
@@ -1,19 +1,0 @@
-{
-  "bindings": [
-    {
-      "authLevel": "function",
-      "name": "req",
-      "type": "httpTrigger",
-      "direction": "in",
-      "methods": [ "get" ]
-    },
-    {
-      "name": "$return",
-      "type": "http",
-      "direction": "out"
-    }
-  ],
-  "disabled": false,
-  "entryPoint": "NCS.DSS.DigitalIdentity.GetDigitalIdentityHttpTrigger.Run",
-  "scriptFile": "..\\bin\\NCS.DSS.DigitalIdentity.dll"
-}
